#1296 complete IgniteSqlFilterClause changes and add tests
- also removes `convertExternalValueToString` from `SchemaMapper`
  and moves it to Filtering use-case as it has nothing to do with
  schema mapping logic.

+++ plugin/ignite-plugin/src/main/scala/org/finos/vuu/feature/ignite/filter/IgniteSqlFilterTreeVisitor.scala
@@ -27,16 +27,16 @@ class IgniteSqlFilterTreeVisitor extends FilterBaseVisitor[IgniteSqlFilterClause
     NeqIgniteSqlFilterClause(ctx.ID().getText, ctx.scalar().getText)
 
   override def visitOperationGt(ctx: OperationGtContext): IgniteSqlFilterClause =
-    RangeIgniteSqlFilterClause(RangeOp.GT)(ctx.ID().getText, ctx.NUMBER().getText.toDouble)
+    RangeIgniteSqlFilterClause(RangeOp.GT)(ctx.ID().getText, ctx.NUMBER().getText)
 
   override def visitOperationGte(ctx: OperationGteContext): IgniteSqlFilterClause =
-    RangeIgniteSqlFilterClause(RangeOp.GTE)(ctx.ID().getText, ctx.NUMBER().getText.toDouble)
+    RangeIgniteSqlFilterClause(RangeOp.GTE)(ctx.ID().getText, ctx.NUMBER().getText)
 
   override def visitOperationLt(ctx: OperationLtContext): IgniteSqlFilterClause =
-    RangeIgniteSqlFilterClause(RangeOp.LT)(ctx.ID().getText, ctx.NUMBER().getText.toDouble)
+    RangeIgniteSqlFilterClause(RangeOp.LT)(ctx.ID().getText, ctx.NUMBER().getText)
 
   override def visitOperationLte(ctx: OperationLteContext): IgniteSqlFilterClause =
-    RangeIgniteSqlFilterClause(RangeOp.LTE)(ctx.ID().getText, ctx.NUMBER().getText.toDouble)
+    RangeIgniteSqlFilterClause(RangeOp.LTE)(ctx.ID().getText, ctx.NUMBER().getText)
 
   override def visitOperationStarts(ctx: OperationStartsContext): IgniteSqlFilterClause =
     StartsIgniteSqlFilterClause(ctx.ID().getText, ctx.STRING().getText)
